+++ NEWS.md
@@ -1,7 +1,6 @@
 # textrecipes (development version)
 
-## Changed behaviour
-
+* `step_word_embeddings()` has been added. Allows for use of pre-trained word embeddings to convert token columns to vectors in a high-dimensional "meaning" space. (@jonthegeek, #20)
 * text2vec have been changed from Imports to Suggests.
 * textfeatures have been changed from Imports to Suggests.
 * step_tfidf() calculations are slightly changed due to flaw in original implementation https://github.com/dselivanov/text2vec/issues/280.